backup_20260415_060429/requirements.txt
python-telegram-bot==20.7
asyncpg
python-dotenv
reportlab
pyaes
pillow
bcrypt
# PTB 20.7 绑定 httpx 0.25.x。若曾单独升级 httpcore 1.x，旧版在同步 main() 里建 AsyncClient 可能触发
# sniffio AsyncLibraryNotFoundError；bot 已用 DeferredHTTPXRequest 推迟到事件循环内再建连。
httpx==0.25.2
httpcore==1.0.9
sniffio==1.3.1
aiohttp
qrcode[pil]
